get and set location and other fixes and deletions

+++ back_end/src/main/java/back/model/Admin.java
@@ -1,7 +1,5 @@
 package back.model;
 
-import back.api.JTHAuth;
-
 import java.util.LinkedHashMap;
 
 public class Admin extends User {
@@ -23,7 +21,7 @@ public class Admin extends User {
 
     public Admin(LinkedHashMap M) {
         super(
-                ((Integer)M.get("id")).intValue(),
+                (Integer) M.get("id"),
                 (String) M.get("email"),
                 (String) M.get("role")
         );
@@ -38,6 +36,4 @@ public class Admin extends User {
     public String getSurname() {
         return surname;
     }
-
-
 }
