+++ back_end/src/main/java/back/model/Provider.java
@@ -19,7 +19,7 @@ public class Provider extends User{
 
     public Provider(LinkedHashMap M){
         super(
-                ((Integer)M.get("id")).intValue(),
+                (Integer) M.get("id"),
                 (String) M.get("email"),
                 (String) M.get("role")
         );
